通过重新组织代码，解决删除单条规则时多次重复渲染的bug
