+++ extension/options_ui/advance.html
@@ -213,6 +213,6 @@
         </section>
       </div>
     </div>
-    <script src="/options_ui/js/advance.js"></script>
+    <script src="/options_ui/js/main.js"></script>
   </body>
 </html>
