+++ extension/options_ui/css/app.css
@@ -211,7 +211,8 @@ font-size: 1.3rem;
   appearance: none;
 }
 
-.notice {
+.notice,
+.warning {
   font-size: 1rem;
   font-weight: 400;
   line-height: 1.5;
