+++ extension/options_ui/index.html
@@ -48,7 +48,7 @@
           <h2>默认侯选项切换 ：</h2>
           <button class="reset_default_domain">重置默认选项</button
           ><span class="placeholder-box"></span>
-          <span class="notice">备注：选中候选项 即 自动完成选项切换</span>
+          <span class="warning">备注：选中候选项 即 自动完成选项切换</span>
           <span class="default-notice"
             >默认配置信息和原理：
             https://github.com/justjavac/ReplaceGoogleCDN/tree/master/extension/options_ui</span
@@ -121,6 +121,6 @@
         </section>
       </div>
     </div>
-    <script src="/options_ui/js/index.js"></script>
+    <script src="/options_ui/js/main.js"></script>
   </body>
 </html>
