+++ extension/options_ui/js/AdvanceBundle/Router/advanceRouter.js
@@ -1,0 +1,5 @@
+import advanceController from "../Controller/advanceController.js";
+
+export default () => {
+  advanceController();
+};
