+++ extension/options_ui/js/AdvanceBundle/app.js
@@ -1,0 +1,5 @@
+import advanceRouter from "./Router/advanceRouter.js";
+
+export default () => {
+  advanceRouter();
+};
