+++ extension/options_ui/js/IndexBundle/Config/defaultDomains.js
@@ -1,0 +1,85 @@
+//domain 与 rules/mirrors 目录下文件名称一一对应
+let defaultDomains = {
+  "ajax.googleapis.com": {
+    id: 1,
+    domain: "ajax.googleapis.com",
+  },
+  "fonts.googleapis.com": {
+    id: 2,
+    domain: "fonts.googleapis.com",
+  },
+  "themes.googleusercontent.com": {
+    id: 3,
+    domain: "themes.googleusercontent.com",
+  },
+  "fonts.gstatic.com": {
+    id: 4,
+    domain: "fonts.gstatic.com",
+  },
+  "gravatar.com": {
+    id: 5,
+    domain: "gravatar.com",
+  },
+  "maxcdn.bootstrapcdn.com/bootstrap": {
+    id: 6,
+    domain: "maxcdn.bootstrapcdn.com",
+  },
+  "code.jquery.com/jquery-(version)(suffix)": {
+    id: 7,
+    domain: "code.jquery.com",
+  },
+  "code.jquery.com/ui/(prefix)(version)(suffix)": {
+    id: 8,
+    domain: "code.jquery.com-ui",
+  },
+  "cdnjs.cloudflare.com": {
+    id: 9,
+    domain: "cdnjs.cloudflare.com",
+  },
+  "cdn.jsdelivr.net": {
+    id: 10,
+    domain: "cdn.jsdelivr.net",
+  },
+  /*
+    "translate.googleapis.com": {
+      id: 11,
+      domain: "translate.googleapis.com",
+    },
+     */
+  "www.gstatic.com": {
+    id: 12,
+    domain: "www.gstatic.com",
+  },
+  "developer.android.com": {
+    id: 13,
+    domain: "developer.android.com",
+  },
+  "developers.google.com": {
+    id: 17,
+    domain: "developers.google.com",
+  },
+  "source.android.com": {
+    id: 18,
+    domain: "source.android.com",
+  },
+  "lh3.googleusercontent.com": {
+    id: 14,
+    domain: "lh3.googleusercontent.com",
+  },
+  "cdn.sstatic.net": {
+    id: 15,
+    domain: "cdn.sstatic.net",
+  },
+  "imgur.com": {
+    id: 16,
+    domain: "imgur.com",
+  },
+  /*
+    "supper-priority-override-rule": {
+      id: 9999,
+      domain: "supper-priority-override",
+    },
+     */
+};
+
+export { defaultDomains };
