+++ extension/options_ui/js/IndexBundle/Router/indexRouter.js
@@ -1,0 +1,4 @@
+import indexController from "../Controller/indexController.js";
+export default () => {
+  indexController();
+};
