+++ extension/options_ui/js/IndexBundle/app.js
@@ -1,0 +1,5 @@
+import indexRouter from "./Router/indexRouter.js";
+
+export default () => {
+  indexRouter();
+};
